consolidate session sender class into sessiontracker
