--- sdk/src/androidTest/java/com/bugsnag/android/LifecycleBreadcrumbLoggerTest.java
@@ -1,12 +1,0 @@
-package com.bugsnag.android;
-
-import android.util.Pair;
-
-import org.junit.Test;
-
-import static junit.framework.Assert.assertEquals;
-import static junit.framework.Assert.assertTrue;
-
-public class LifecycleBreadcrumbLoggerTest {
-
-}
